add `create_if_not_exists` and refactor `Read.py`
Now uses `os.path.join` to add the json folder into the path if its not in the name which should improve things a lot.
also added `create_if_not_exists` with an overload because a function may not want it to be created when reading a json file, in that case it returns `None` if the flag is False.